requirements.txt
fastapi
uvicorn
pyjwt
python-jose
passlib
python-dotenv
python-multipart
bcrypt
requests
websockets
pyOpenSSL 